Fix INCBIN/BIN unquoted filenames
asm6's documentation states the following syntaxes are allowed:

INCBIN foo.bin
INCBIN foo.bin, $400
INCBIN foo.bin, $200, $2000

The 1st example works, but the 2nd and 3rd fail with a vague error of
"Can't open file."  The filename attempted to be opened isn't shown, so
the user has no idea what's wrong (had it been, I suspect this bug would
have been fixed a lot sooner).

However, if you put the filename part in double-quotes, all 3 forms work
correctly.

The bug: incbin() calls getfilename() (the only piece of code that uses
the getfilename() function).  getfilename() calls getword() with the 3rd
argument (mcheck) being 0.  An mcheck of 0 causes the function to *not*
call strtok() against "mathy letters", meaning it acts like strchr() but
with multiple letters that cause it to stop.  So in the case of the
string "foo.bin, $200, $2000", the string "foo.bin," is returned and
treated as the filename (note the trailing comma).

The fix: extend getword() to support an mcheck of 2, which calls
strtok(dst,",") and thus only gets the filename portion; also consider
that the size/offset arguments to INCBIN are comma-delimited.

I could have simply changed getword() to use an mcheck value of 1, but
that would allow syntactically-incorrect things to work, such as
"INCBIN foo.bin^ $200".

Finally, document all of this in readme.txt, as well as add some basic
test case code in test_cases/incbin.  (Note there is no simple way way
to implement these tests, and I dislike "test suite frameworks," so just
refer to the comments in the test code for now.)