Makefile fixups (Cygwin vs. Windows vs. Linux)
- Leverage $CC, as it's set by GNU make: https://www.gnu.org/software/make/manual/html_node/Implicit-Variables.html
- Detect Cygwin vs. native Windows vs. Linux better (and without shell)
- Deal with rm vs. del (native Windows) better

The logic here can be confusing so I'll explain it with a chart:

| Environment       | OS         | OSTYPE      |
| ----------------- | ---------- | ----------- |
| Cygwin on Windows | Windows_NT | cygwin      |
| Windows cmd.exe   | Windows_NT | not set     |
| FreeBSD           | not set    | freebsdXX.Y |
| Linux             | not set    | linux-gnu   |

MSYS2/MSYS/MinGW (on Windows) may add complications to this logic; I
haven't tested them, but hope they set $OSTYPE.

And the BSDs require GNU make (gmake) installed to use any of this (i.e.
BSD make is not compatible with any of this, but that's expected).

Regarding $CC:

Assuming gcc isn't exactly safe everywhere nowadays.  FreeBSD since 11.x
has used clang instead of gcc (there is no gcc command), while "cc" is
is a symlink to "clang", so it works fine.  Thus, just rely on what
GNU make defines $CC as by default; else we can override it elsewhere.