Code safety: use snprintf instead of sprintf
Note that while here, we have to deal with maximum pathname lengths.
The underlying asm6 functions call their variables "filename", but
most of these functions are relying on the "outputfilename" global,
which can in fact be a full pathname if specified.  We opt to use
the POSIX standard PATH_MAX #define and hope for the best.

References:

* https://eklitzke.org/path-max-is-tricky
* https://stackoverflow.com/a/65174437